+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - Added `ignoreHelpers` option to skip automatic lookup/bundling of helpers
+- Added `precompileOptions` to pass options to handlebars precompile
+- Your feature here!
 
 ## [1.5.0] - 2017-04-20
 
@@ -19,7 +21,6 @@
 - Fixed resolving relative helpers on first pass when helper directories are given
 - Added `ignorePartials` option to skip automatic lookup/bundling of partials
 - Added `compat` option to enable Mustache lookup compatibility.
-- Your feature here!
 - Added `config` option to query so that configs can be specified in webpack
   config object or the loader query. Defaults to `handlebarsLoader`
 - Added `partialResolver` config option to override the default partial